src/requirements.txt
Flask==3.0.3
psycopg2==2.9.9
flask-socketio==5.3.1
python-dotenv==1.0.1
pytest==8.3.3